# Conflicts:
#	backend/src/main/java/fr/projet/kitcinq/model/CourseEntity.java
#	backend/src/main/java/fr/projet/kitcinq/model/FormationEntity.java
#	backend/src/main/java/fr/projet/kitcinq/model/ProfessorEntity.java
#	backend/src/main/java/fr/projet/kitcinq/model/StudentCourseEntity.java
#	backend/src/main/java/fr/projet/kitcinq/model/StudentCourseId.java
#	backend/src/main/java/fr/projet/kitcinq/student/studentcall/StudentCallController.java
#	backend/src/main/java/fr/projet/kitcinq/student/studentcall/StudentCourseRepository.java
#	backend/src/main/java/fr/projet/kitcinq/user/SecurityConfig.java
#	backend/src/main/java/fr/projet/kitcinq/user/UserController.java

+++ backend/src/main/java/fr/projet/kitcinq/student/studentcall/StudentCallService.java
@@ -1,12 +1,7 @@
 package fr.projet.kitcinq.student.studentcall;
 
 public interface StudentCallService {
-    
-    void launch(int courseId);
-    
-    void setStudentPresent(int courseId, int studentId);
-    
-    void setStudentAbsent(int courseId, int studentId);
-    
-    
+    void launch(long courseId);
+    void setStudentPresent(long courseId, long studentId);
+    void setStudentAbsent(long courseId, long studentId);
 }
